requirements.txt
Django==1.5
MySQL-python==1.2.4
South==0.7.6
argparse==1.2.1
django-simple-rest==1.4.1
mimeparse==0.1.3
wadofstuff-django-serializers==1.1.0
wsgiref==0.1.2
